requirements.txt
black==26.3.1
isort==6.0.1
pyTelegramBotAPI==4.26.0
RapidFuzz==3.0.0
toml==0.10.2
captcha==0.7.1
pillow==11.1.0
